Steuerung_Versuchsstand/Raspberry/requirements.txt
customtkinter==4.6.3
Flask==1.1.2
matplotlib==3.7.2
mlxtend==0.23.0
numpy==1.24.2
numpy==1.19.5
opencv_python==4.8.0.74
pandas==2.0.3
picamera2==0.3.12
Pillow==8.1.2
Pillow==10.1.0
pyserial==3.5b0
Requests==2.31.0
scikit_learn==1.3.2
tensorboard==2.11.2
tensorflow==2.11.0
tensorflow_cpu_aws==2.11.0
tflite_runtime==2.11.0
tflite_runtime==2.5.0.post1
